Basic DB and partial Save functionality

+++ app/src/main/java/com/projectdolphin/data/model/Class.java
@@ -9,10 +9,10 @@ import java.util.List;
  */
 public class Class extends GradedDBItem {
 
-    public Class(long timeSpentMillis, double grade, double weight, String title, List<Integer> categoryIDs) {
+    public Class(long timeSpentMillis, double grade, double weight, String title, List<Long> categoryIDs) {
         this(-1, timeSpentMillis, grade, weight, title, categoryIDs);
     }
-    public Class(long DB_ID, long timeSpentMillis, double grade, double weight, String title, List<Integer> categoryIDs) {
+    public Class(long DB_ID, long timeSpentMillis, double grade, double weight, String title, List<Long> categoryIDs) {
         super(DB_ID, timeSpentMillis, grade, weight, title);
         this.categoryIDs = categoryIDs;
     }
@@ -22,12 +22,12 @@ public class Class extends GradedDBItem {
         return String.format("%.0f hours", getWeight());
     }
 
-    public List<Integer> getCategoryIDs() {
+    public List<Long> getCategoryIDs() {
         return categoryIDs;
     }
-    public void setCategoryIDs(List<Integer> categoryIDs) {
+    public void setCategoryIDs(List<Long> categoryIDs) {
         this.categoryIDs = categoryIDs;
     }
 
-    private List<Integer> categoryIDs;
+    private List<Long> categoryIDs;
 }
